US-2 Implement ticket ordering

+++ src/main/scala/config/Discount.scala
@@ -4,11 +4,11 @@ package config
 import zio.config.*
 import zio.config.ConfigDescriptor.*
 
-final case class Discount(afterDays: Int, percentage: Int)
+final case class Discount(afterDays: Long, percentage: Int)
 
 object Discount {
-  val afterDays: ConfigDescriptor[Int] =
-    int("after-days").describe("After how many days of a show's opening day should a discount be applied")
+  val afterDays: ConfigDescriptor[Long] =
+    long("after-days").describe("After how many days of a show's opening day should a discount be applied")
 
   val percentage: ConfigDescriptor[Int] =
     int("percentage").describe("Discount percentage to apply")
